+++ plugins/VolumeOfInterest/VolumeOfInterest.h
@@ -41,8 +41,8 @@ protected slots:
   void defineVOI(SelectionHandler::MultiSelection msel);
   void cancelVOI();
 
-  void setBorderFrom(int pos/*nm*/, vtkPVSliceView::VIEW_PLANE plane);
-  void setBorderTo(int pos/*nm*/, vtkPVSliceView::VIEW_PLANE plane);
+  void setBorderFrom(double pos/*nm*/, vtkPVSliceView::VIEW_PLANE plane);
+  void setBorderTo(double pos/*nm*/, vtkPVSliceView::VIEW_PLANE plane);
 
 private:
   void buildVOIs();
